fastapi/requirements.txt
fastapi
fastapi_utils
uvicorn[standard]
gunicorn
alembic
redis
psycopg2-binary
pydantic
pydantic_settings
pytest
python-dotenv
python-jose[cryptography]
python-multipart
jwt
bcrypt
simplejson
requests
pillow
httpx
email-validator
